requirements.txt
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
pycodestyle==2.8.0
pyparsing==3.0.9
pytesseract==0.3.9
python-decouple==3.6
pytube==12.1.0
requests==2.27.1
toml==0.10.2
urllib3==1.26.9
